image_upload/requirements.txt
azure-ai-vision-imageanalysis==1.0.0b2
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1